Split functionality into seperate modules for better maintainability, added bucketing functionality, removed old single script
